+++ .npmignore
@@ -1,2 +1,2 @@
-test/*ml
+test
 .github
